// apps/web/src/types/generated/ClosedReason.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HolidayKind } from "./HolidayKind";

/**
 * 休市原因。
 */
export type ClosedReason = "Weekend" | {
  "OfficialHoliday": { citation_id: string };
} | { "SimulatedHoliday": HolidayKind };